pyproject.toml
[project]
name = "jira-track"
version = "0.1.0"
description = ""
authors = [
    {name = "Micke113",email = "michaellagarde113@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5,<3.0.0",
    "textual>=8.2.3",
    "xlsxwriter>=3.2.9,<4.0.0",
]

[project.scripts]
jira-track = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = [
    "src",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
